backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
firebase-admin==6.2.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
transformers==4.35.0
torch
torchvision
tiktoken==0.5.1
sentence-transformers==2.2.2
